Test TensorTypeSet instead of autograd_meta_ for variable-ness.
By the current autograd_meta_ <=> type_set_ invariant (now explicitly documented
in the right place!), these are equivalent.  But when I introduce null
autograd_meta_ optimization, they won't be equivalent anymore: TensorTypeSet is
going to give me the right information no matter what.

In the long run, this patch will be a wash, because everything will "be a variable"
in the long term.  But I am making this change now to make sure that the invariant
actually holds.

Signed-off-by: Edward Z. Yang <ezyang@fb.com>

ghstack-source-id: 4db17850bacea6f84e2e3058146aed9bc1827e06
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28543