[pytorch] Make FileStore not segfault with concurrent accesses.
Pull Request resolved: https://github.com/pytorch/pytorch/pull/28812

FileStore wasn't thread-safe. Maybe it wasn't necessarily intended
as such, but it's not documented as such, and we already have
unittests that use this store from multiple threads, and
occasionally crash as a result.

We can address this by simply throwing a small-scope std::mutex
around accesses to the std::unordered_map<> cache_.
ghstack-source-id: 92787892

Differential Revision: [D18187821](https://our.internmc.facebook.com/intern/diff/D18187821/)